+++ src/middlewares/index.ts
@@ -1,3 +1,4 @@
+import chalk from 'chalk';
 import * as _ from 'lodash';
 import { Locals } from "../types/common.types";
 import CRUD from '../utils/crud';
@@ -17,7 +18,7 @@ const _IterateResponse = (_req, res, next) => {
   const store = routeConfig.store || {};
 
   if (!Array.isArray(locals.data)) {
-    console.error("To use _IterateResponse method the data must be of type Array");
+    console.error(chalk.red("To use ") + chalk.yellowBright("_IterateResponse") + chalk.red(" method the data must be of type Array"));
     return next();
   }
 
@@ -38,7 +39,7 @@ const _IterateRoutes = (req, res, next) => {
   const store = routeConfig.store || {};
 
   if (!Array.isArray(locals.data)) {
-    console.error("To use _IterateRoutes method the data must be of type Array");
+    console.error(chalk.red("To use ") + chalk.yellowBright("_IterateRoutes") + chalk.red(" method the data must be of type Array"));
     return next();
   }
 
@@ -63,7 +64,7 @@ const _CrudOperation = (req, res, next) => {
   const store = routeConfig.store || {};
 
   if (!(_.isArray(locals.data) && locals.data.every(d => _.isPlainObject(d)))) {
-    console.error("To use _CurdResponse method the data must be of type Array of objects");
+    console.error(chalk.red("To use ") + chalk.yellowBright("_CurdResponse") + chalk.red(" method the data must be of type Array"));
     return next();
   }
   if (!store[storeKey]) store[storeKey] = _.cloneDeep(locals.data);
@@ -85,9 +86,16 @@ const _CrudOperation = (req, res, next) => {
 const _AdvancedSearch = (req, res, next) => {
   const method = req.method;
   const locals = res.locals as Locals;
+  
+  if (!(_.isArray(locals.data) && locals.data.every(d => _.isPlainObject(d)))) {
+    console.error(chalk.red("To use ") + chalk.yellowBright("_AdvancedSearch") + chalk.red(" method the data must be of type Array"));
+    return next();
+  }
+
   if (method.toLowerCase() === 'get') {
     locals.data = CRUD.search(req, res, locals.data);
   }
+  
   next();
 }
 const _FetchTillData = (_req, res, next) => {