+++ src/middlewares/initialMiddleware.ts
@@ -25,7 +25,7 @@ export default (routePath: string, config: ValidTypes.Config, getDb: (routePath?
 
       if (_.isPlainObject(routeConfig.headers) && !_.isEmpty(routeConfig.headers)) {
         Object.entries(routeConfig.headers as object).forEach(([headerName, value]) => {
-          res.set(headerName, value);
+          res.setHeader(headerName, value);
         })
       }